New (updated) .csv file with correct data
note that I also added sacks, pass attempts, pass completions, passing interceptions, defensive interceptions. Also note that the teams are based on which team that specific player plays for. Example: A.Rodgers plays for NYJ his team will be NYJ, opponent_team will be SF. But K.Juszczyk plays for SF, his team will be SF, opponent_team will be NYJ. This also changes the game_id to reflect this. This could be a potential problem later that will need to be fixed.